Extract exception classes into bot/exceptions.py
Moves all exception classes (SkipPage, HasFragment, IsDisambigPage,
InstanceOfForbidden, IsRedirectWithItemPage, IsRedirectPage, PageGone,
SettingsReloadedException) into their own module with no side effects.

common.py re-exports them for backward compatibility.
